requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
aioimaplib>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0

# Cloud (optional)
# asyncpg>=0.29.0
# cryptography>=41.0.0

# LLM providers (optional)
# openai>=1.10.0
# anthropic>=0.18.0

# Dev
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
